admin_client/client_gui/requirements.txt
requests==2.23.0
PyQt5==5.14.2
python-dotenv==0.13.0
